Add logging handlers only when necessary
Trying to debug using this library under a root-only filesystem
uncovered that the handlers will initialize themselves even if they
aren't active in a logger. This means that the file handlers will for
instance, try to write to the log file in question even though that's an
impossible situation. This re-structure only adds the handler to a the
logging dictconf when actually being utilized.